Splitting the UnicodeTest.io file into unit tests and a helper script
UnicodeTest.io was doing double duty. It held both the unit tests, and acted as
a script which could be run to get Io to print certain test strings and then
exit. Half of the unit tests launch a separate Io process which runs the helper
script.

This patch leaves the unit tests in the UnicodeTest.io file, and splits the
helper executable script out into its own file.