Only include IoVersion.h in .c file that uses it
IoVersion.h defines a single macro that has the Io version number. It
was included in IoState.h to make the define global, but only IoSystem.c
needs the version number.

Reducing the dependencies on the version number helps minimize the
amount that has to be recompiled on version number change.