Various changes to support a cleaner build and install on Win32.
Update the Makefile system to allow specification of a
library/dll prefix, as on Win32 there is normally no
'lib' prefix.

Various minor changes to support the non C99 status of VC8.

Various changes to the new PHash code to work under Windows,
it seems that the recent refactor introduced some API
errors, the PHash functions were still being marked with the
BASEKIT_API macro.

Change the implementation of the "System installPrefix" on Win32
only to calculate the install location from the binary location
instead of fixing it at compile time. This will allow the Win32
version to be installed anywhere.